+++ docs/lib/angular-ui-router/release/angular-ui-router.js
@@ -4,7 +4,7 @@
  *         This causes it to be incompatible with plugins that depend on @uirouter/core.
  *         We recommend switching to the ui-router-core.js and ui-router-angularjs.js bundles instead.
  *         For more information, see https://ui-router.github.io/blog/uirouter-for-angularjs-umd-bundles
- * @version v1.0.7
+ * @version v1.0.12
  * @link https://ui-router.github.io
  * @license MIT License, http://www.opensource.org/licenses/MIT
  */
@@ -499,15 +499,16 @@ var services = {
  * @module common
  */
 /** for typedoc */
-var w = typeof window === 'undefined' ? {} : window;
-var angular$1 = w.angular || {};
+var root = (typeof self === 'object' && self.self === self && self) ||
+    (typeof global === 'object' && global.global === global && global) || undefined;
+var angular$1 = root.angular || {};
 var fromJson = angular$1.fromJson || JSON.parse.bind(JSON);
 var toJson = angular$1.toJson || JSON.stringify.bind(JSON);
 var forEach = angular$1.forEach || _forEach;
 var extend = Object.assign || _extend;
 var equals = angular$1.equals || _equals;
 function identity(x) { return x; }
-function noop$1() { }
+function noop() { }
 /**
  * Builds proxy functions on the `to` object which pass through to the `from` object.
  *
@@ -1070,7 +1071,6 @@ var Queue = /** @class */ (function () {
  * @coreapi
  * @module transition
  */ /** for typedoc */
-"use strict";
 
 (function (RejectType) {
     RejectType[RejectType["SUPERSEDED"] = 2] = "SUPERSEDED";
@@ -1203,7 +1203,9 @@ function normalizedCat(input) {
     return isNumber(input) ? exports.Category[input] : exports.Category[exports.Category[input]];
 }
 /** @hidden */
-var consoletable = isFunction(console.table) ? console.table.bind(console) : console.log.bind(console);
+var consoleLog = Function.prototype.bind.call(console.log, console);
+/** @hidden */
+var consoletable = isFunction(console.table) ? console.table.bind(console) : consoleLog.bind(console);
 /**
  * Trace categories Enum
  *
@@ -1348,10 +1350,10 @@ var Trace = /** @class */ (function () {
         if (!this.enabled(exports.Category.VIEWCONFIG))
             return;
         var mapping = pairs.map(function (_a) {
-            var uiViewData = _a[0], config = _a[1];
-            var uiView = uiViewData.$type + ":" + uiViewData.fqn;
-            var view = config && config.viewDecl.$context.name + ": " + config.viewDecl.$name + " (" + config.viewDecl.$type + ")";
-            return { 'ui-view fqn': uiView, 'state: view name': view };
+            var uiView = _a.uiView, viewConfig = _a.viewConfig;
+            var uiv = uiView && uiView.fqn;
+            var cfg = viewConfig && viewConfig.viewDecl.$context.name + ": " + viewConfig.viewDecl.$name;
+            return { 'ui-view fqn': uiv, 'state: view name': cfg };
         }).sort(function (a, b) { return a['ui-view fqn'].localeCompare(b['ui-view fqn']); });
         consoletable(mapping);
     };
@@ -1542,7 +1544,7 @@ var TargetState = /** @class */ (function () {
  */
 /** for typedoc */
 var defaultOptions = {
-    current: noop$1,
+    current: noop,
     transition: null,
     traceData: {},
     bind: null,
@@ -1599,6 +1601,11 @@ var TransitionHook = /** @class */ (function () {
         catch (err) {
             // If callback throws (synchronously)
             return handleError(Rejection.normalize(err));
+        }
+        finally {
+            if (hook.invokeLimit && ++hook.invokeCount >= hook.invokeLimit) {
+                hook.deregister();
+            }
         }
     };
     /**
@@ -1781,15 +1788,18 @@ function matchState(state, criterion) {
  * The registration data for a registered transition hook
  */
 var RegisteredHook = /** @class */ (function () {
-    function RegisteredHook(tranSvc, eventType, callback, matchCriteria, options) {
+    function RegisteredHook(tranSvc, eventType, callback, matchCriteria, removeHookFromRegistry, options) {
         if (options === void 0) { options = {}; }
         this.tranSvc = tranSvc;
         this.eventType = eventType;
         this.callback = callback;
         this.matchCriteria = matchCriteria;
+        this.removeHookFromRegistry = removeHookFromRegistry;
+        this.invokeCount = 0;
+        this._deregistered = false;
         this.priority = options.priority || 0;
         this.bind = options.bind || null;
-        this._deregistered = false;
+        this.invokeLimit = options.invokeLimit;
     }
     /**
      * Gets the matching [[PathNode]]s
@@ -1870,6 +1880,10 @@ var RegisteredHook = /** @class */ (function () {
         var allMatched = values(matches).every(identity);
         return allMatched ? matches : null;
     };
+    RegisteredHook.prototype.deregister = function () {
+        this.removeHookFromRegistry(this);
+        this._deregistered = true;
+    };
     return RegisteredHook;
 }());
 /** @hidden Return a registration function of the requested type. */
@@ -1877,16 +1891,14 @@ function makeEvent(registry, transitionService, eventType) {
     // Create the object which holds the registered transition hooks.
     var _registeredHooks = registry._registeredHooks = (registry._registeredHooks || {});
     var hooks = _registeredHooks[eventType.name] = [];
+    var removeHookFn = removeFrom(hooks);
     // Create hook registration function on the IHookRegistry for the event
     registry[eventType.name] = hookRegistrationFn;
     function hookRegistrationFn(matchObject, callback, options) {
         if (options === void 0) { options = {}; }
-        var registeredHook = new RegisteredHook(transitionService, eventType, callback, matchObject, options);
+        var registeredHook = new RegisteredHook(transitionService, eventType, callback, matchObject, removeHookFn, options);
         hooks.push(registeredHook);
-        return function deregisterEventHook() {
-            registeredHook._deregistered = true;
-            removeFrom(hooks)(registeredHook);
-        };
+        return registeredHook.deregister.bind(registeredHook);
     }
     return hookRegistrationFn;
 }
@@ -3085,9 +3097,9 @@ var Transition = /** @class */ (function () {
      * #### Example:
      * ```js
      * .onEnter({ to: 'foo.bar' }, trans => {
-     *   // returns result of `foo` state's `data` resolve
-     *   // even though `foo.bar` also has a `data` resolve
-     *   var fooData = trans.injector('foo').get('data');
+     *   // returns result of `foo` state's `myResolve` resolve
+     *   // even though `foo.bar` also has a `myResolve` resolve
+     *   var fooData = trans.injector('foo').get('myResolve');
      * });
      * ```
      *
@@ -3096,8 +3108,8 @@ var Transition = /** @class */ (function () {
      * #### Example:
      * ```js
      * .onExit({ exiting: 'foo.bar' }, trans => {
-     *   // Gets the resolve value of `data` from the exiting state.
-     *   var fooData = trans.injector(null, 'foo.bar').get('data');
+     *   // Gets the resolve value of `myResolve` from the state being exited
+     *   var fooData = trans.injector(null, 'from').get('myResolve');
      * });
      * ```
      *
@@ -3152,6 +3164,18 @@ var Transition = /** @class */ (function () {
     };
     /**
      * Dynamically adds a new [[Resolvable]] (i.e., [[StateDeclaration.resolve]]) to this transition.
+     *
+     * Allows a transition hook to dynamically add a Resolvable to this Transition.
+     *
+     * Use the [[Transition.injector]] to retrieve the resolved data in subsequent hooks ([[UIInjector.get]]).
+     *
+     * If a `state` argument is provided, the Resolvable is processed when that state is being entered.
+     * If no `state` is provided then the root state is used.
+     * If the given `state` has already been entered, the Resolvable is processed when any child state is entered.
+     * If no child states will be entered, the Resolvable is processed during the `onFinish` phase of the Transition.
+     *
+     * The `state` argument also scopes the resolved data.
+     * The resolved data is available from the injector for that `state` and any children states.
      *
      * #### Example:
      * ```js
@@ -3599,7 +3623,7 @@ var beforeAfterSubstr = function (char) { return function (str) {
     return [str.substr(0, idx), str.substr(idx + 1)];
 }; };
 var hostRegex = new RegExp('^(?:[a-z]+:)?//[^/]+/');
-var stripFile = function (str) { return str.replace(/\/[^/]*$/, ''); };
+var stripLastPathElement = function (str) { return str.replace(/\/[^/]*$/, ''); };
 var splitHash = beforeAfterSubstr("#");
 var splitQuery = beforeAfterSubstr("?");
 var splitEqual = beforeAfterSubstr("=");
@@ -3837,8 +3861,8 @@ var StateParams = /** @class */ (function () {
 var parseUrl = function (url) {
     if (!isString(url))
         return false;
-    var root = url.charAt(0) === '^';
-    return { val: root ? url.substring(1) : url, root: root };
+    var root$$1 = url.charAt(0) === '^';
+    return { val: root$$1 ? url.substring(1) : url, root: root$$1 };
 };
 function nameBuilder(state) {
     return state.name;
@@ -3853,7 +3877,7 @@ function dataBuilder(state) {
     }
     return state.data;
 }
-var getUrlBuilder = function ($urlMatcherFactoryProvider, root) {
+var getUrlBuilder = function ($urlMatcherFactoryProvider, root$$1) {
     return function urlBuilder(state) {
         var stateDec = state;
         // For future states, i.e., states whose name ends with `.**`,
@@ -3874,7 +3898,7 @@ var getUrlBuilder = function ($urlMatcherFactoryProvider, root) {
             return null;
         if (!$urlMatcherFactoryProvider.isMatcher(url))
             throw new Error("Invalid url '" + url + "' in state '" + state + "'");
-        return (parsed && parsed.root) ? url : ((parent && parent.navigable) || root()).url.append(url);
+        return (parsed && parsed.root) ? url : ((parent && parent.navigable) || root$$1()).url.append(url);
     };
 };
 var getNavigableBuilder = function (isRoot) {
@@ -4002,12 +4026,12 @@ var StateBuilder = /** @class */ (function () {
     function StateBuilder(matcher, urlMatcherFactory) {
         this.matcher = matcher;
         var self = this;
-        var root = function () { return matcher.find(""); };
+        var root$$1 = function () { return matcher.find(""); };
         var isRoot = function (state) { return state.name === ""; };
         function parentBuilder(state) {
             if (isRoot(state))
                 return null;
-            return matcher.find(self.parentName(state)) || root();
+            return matcher.find(self.parentName(state)) || root$$1();
         }
         this.builders = {
             name: [nameBuilder],
@@ -4015,7 +4039,7 @@ var StateBuilder = /** @class */ (function () {
             parent: [parentBuilder],
             data: [dataBuilder],
             // Build a URLMatcher if necessary, either via a relative or absolute URL
-            url: [getUrlBuilder(urlMatcherFactory, root)],
+            url: [getUrlBuilder(urlMatcherFactory, root$$1)],
             // Keep track of the closest ancestor state that has a URL (i.e. is navigable)
             navigable: [getNavigableBuilder(isRoot)],
             params: [getParamsBuilder(urlMatcherFactory.paramFactory)],
@@ -4067,7 +4091,7 @@ var StateBuilder = /** @class */ (function () {
         for (var key in builders) {
             if (!builders.hasOwnProperty(key))
                 continue;
-            var chain = builders[key].reduce(function (parentFn, step) { return function (_state) { return step(_state, parentFn); }; }, noop$1);
+            var chain = builders[key].reduce(function (parentFn, step) { return function (_state) { return step(_state, parentFn); }; }, noop);
             state[key] = chain(state);
         }
         return state;
@@ -5231,7 +5255,7 @@ function appendBasePath(url, isHtml5, absolute, baseHref) {
     if (baseHref === '/')
         return url;
     if (isHtml5)
-        return stripFile(baseHref) + url;
+        return stripLastPathElement(baseHref) + url;
     if (absolute)
         return baseHref.slice(1) + url;
     return url;
@@ -5396,10 +5420,10 @@ var UrlRouter = /** @class */ (function () {
     UrlRouter.prototype.update = function (read) {
         var $url = this._router.locationService;
         if (read) {
-            this.location = $url.path();
+            this.location = $url.url();
             return;
         }
-        if ($url.path() === this.location)
+        if ($url.url() === this.location)
             return;
         $url.url(this.location, true);
     };
@@ -5552,11 +5576,16 @@ var ViewService = /** @class */ (function () {
         this._uiViews = [];
         this._viewConfigs = [];
         this._viewConfigFactories = {};
+        this._listeners = [];
         this._pluginapi = {
             _rootViewContext: this._rootViewContext.bind(this),
             _viewConfigFactory: this._viewConfigFactory.bind(this),
             _registeredUIViews: function () { return _this._uiViews; },
             _activeViewConfigs: function () { return _this._viewConfigs; },
+            _onSync: function (listener) {
+                _this._listeners.push(listener);
+                return function () { return removeFrom(_this._listeners, listener); };
+            },
         };
     }
     ViewService.prototype._rootViewContext = function (context) {
@@ -5618,19 +5647,24 @@ var ViewService = /** @class */ (function () {
                 // console.log(`Multiple matching view configs for ${uiView.fqn}`, matchingConfigs);
                 matchingConfigs.sort(depthCompare(viewConfigDepth, -1)); // descending
             }
-            return [uiView, matchingConfigs[0]];
+            return { uiView: uiView, viewConfig: matchingConfigs[0] };
         };
-        var configureUIView = function (_a) {
-            var uiView = _a[0], viewConfig = _a[1];
+        var configureUIView = function (tuple) {
             // If a parent ui-view is reconfigured, it could destroy child ui-views.
             // Before configuring a child ui-view, make sure it's still in the active uiViews array.
-            if (_this._uiViews.indexOf(uiView) !== -1)
-                uiView.configUpdated(viewConfig);
+            if (_this._uiViews.indexOf(tuple.uiView) !== -1)
+                tuple.uiView.configUpdated(tuple.viewConfig);
         };
         // Sort views by FQN and state depth. Process uiviews nearest the root first.
-        var pairs$$1 = this._uiViews.sort(depthCompare(uiViewDepth, 1)).map(matchingConfigPair);
-        trace.traceViewSync(pairs$$1);
-        pairs$$1.forEach(configureUIView);
+        var uiViewTuples = this._uiViews.sort(depthCompare(uiViewDepth, 1)).map(matchingConfigPair);
+        var matchedViewConfigs = uiViewTuples.map(function (tuple) { return tuple.viewConfig; });
+        var unmatchedConfigTuples = this._viewConfigs
+            .filter(function (config) { return inArray(matchedViewConfigs, config); })
+            .map(function (viewConfig) { return ({ uiView: undefined, viewConfig: viewConfig }); });
+        var allTuples = uiViewTuples.concat(unmatchedConfigTuples);
+        uiViewTuples.forEach(configureUIView);
+        this._listeners.forEach(function (cb) { return cb(allTuples); });
+        trace.traceViewSync(allTuples);
     };
     
     /**
@@ -5843,7 +5877,7 @@ var UIRouterGlobals = /** @class */ (function () {
  */ /** */
 /** @hidden */
 var makeStub = function (keys) {
-    return keys.reduce(function (acc, key) { return (acc[key] = notImplemented(key), acc); }, { dispose: noop$1 });
+    return keys.reduce(function (acc, key) { return (acc[key] = notImplemented(key), acc); }, { dispose: noop });
 };
 /** @hidden */ var locationServicesFns = ["url", "path", "search", "hash", "onChange"];
 /** @hidden */ var locationConfigFns = ["port", "protocol", "host", "baseHref", "html5Mode", "hashPrefix"];
@@ -6185,10 +6219,11 @@ var registerOnEnterHook = function (transitionService) {
 
 /** @module hooks */
 /** for typedoc */
+var RESOLVE_HOOK_PRIORITY = 1000;
 /**
  * A [[TransitionHookFn]] which resolves all EAGER Resolvables in the To Path
  *
- * Registered using `transitionService.onStart({}, eagerResolvePath);`
+ * Registered using `transitionService.onStart({}, eagerResolvePath, { priority: 1000 });`
  *
  * When a Transition starts, this hook resolves all the EAGER Resolvables, which the transition then waits for.
  *
@@ -6197,15 +6232,15 @@ var registerOnEnterHook = function (transitionService) {
 var eagerResolvePath = function (trans) {
     return new ResolveContext(trans.treeChanges().to)
         .resolvePath("EAGER", trans)
-        .then(noop$1);
+        .then(noop);
 };
 var registerEagerResolvePath = function (transitionService) {
-    return transitionService.onStart({}, eagerResolvePath, { priority: 1000 });
+    return transitionService.onStart({}, eagerResolvePath, { priority: RESOLVE_HOOK_PRIORITY });
 };
 /**
  * A [[TransitionHookFn]] which resolves all LAZY Resolvables for the state (and all its ancestors) in the To Path
  *
- * Registered using `transitionService.onEnter({ entering: () => true }, lazyResolveState);`
+ * Registered using `transitionService.onEnter({ entering: () => true }, lazyResolveState, { priority: 1000 });`
  *
  * When a State is being entered, this hook resolves all the Resolvables for this state, which the transition then waits for.
  *
@@ -6215,10 +6250,28 @@ var lazyResolveState = function (trans, state) {
     return new ResolveContext(trans.treeChanges().to)
         .subContext(state.$$state())
         .resolvePath("LAZY", trans)
-        .then(noop$1);
+        .then(noop);
 };
 var registerLazyResolveState = function (transitionService) {
-    return transitionService.onEnter({ entering: val(true) }, lazyResolveState, { priority: 1000 });
+    return transitionService.onEnter({ entering: val(true) }, lazyResolveState, { priority: RESOLVE_HOOK_PRIORITY });
+};
+/**
+ * A [[TransitionHookFn]] which resolves any dynamically added (LAZY or EAGER) Resolvables.
+ *
+ * Registered using `transitionService.onFinish({}, eagerResolvePath, { priority: 1000 });`
+ *
+ * After all entering states have been entered, this hook resolves any remaining Resolvables.
+ * These are typically dynamic resolves which were added by some Transition Hook using [[Transition.addResolvable]].
+ *
+ * See [[StateDeclaration.resolve]]
+ */
+var resolveRemaining = function (trans) {
+    return new ResolveContext(trans.treeChanges().to)
+        .resolvePath("LAZY", trans)
+        .then(noop);
+};
+var registerResolveRemaining = function (transitionService) {
+    return transitionService.onFinish({}, resolveRemaining, { priority: RESOLVE_HOOK_PRIORITY });
 };
 
 /** @module hooks */ /** for typedoc */
@@ -6235,7 +6288,7 @@ var loadEnteringViews = function (transition) {
     var enteringViews = transition.views("entering");
     if (!enteringViews.length)
         return;
-    return $q.all(enteringViews.map(function (view) { return $q.when(view.load()); })).then(noop$1);
+    return $q.all(enteringViews.map(function (view) { return $q.when(view.load()); })).then(noop);
 };
 var registerLoadEnteringViews = function (transitionService) {
     return transitionService.onFinish({}, loadEnteringViews);
@@ -6500,7 +6553,7 @@ var defaultTransOpts = {
     reload: false,
     custom: {},
     current: function () { return null; },
-    source: "unknown"
+    source: "unknown",
 };
 /**
  * This class provides services related to Transitions.
@@ -6605,7 +6658,7 @@ var TransitionService = /** @class */ (function () {
         var TH = TransitionHook;
         var paths = this._criteriaPaths;
         var NORMAL_SORT = false, REVERSE_SORT = true;
-        var ASYNCHRONOUS = false, SYNCHRONOUS = true;
+        var SYNCHRONOUS = true;
         this._defineEvent("onCreate", Phase.CREATE, 0, paths.to, NORMAL_SORT, TH.LOG_REJECTED_RESULT, TH.THROW_ERROR, SYNCHRONOUS);
         this._defineEvent("onBefore", Phase.BEFORE, 0, paths.to);
         this._defineEvent("onStart", Phase.RUN, 0, paths.to);
@@ -6635,8 +6688,7 @@ var TransitionService = /** @class */ (function () {
         this._eventTypes.push(eventType);
         makeEvent(this, this, eventType);
     };
-    
-    /** @hidden */
+    /** @hidden */ // tslint:disable-next-line
     TransitionService.prototype._getEvents = function (phase) {
         var transitionHookTypes = isDefined(phase) ?
             this._eventTypes.filter(function (type) { return type.hookPhase === phase; }) :
@@ -6662,7 +6714,7 @@ var TransitionService = /** @class */ (function () {
     TransitionService.prototype._definePathType = function (name, hookScope) {
         this._criteriaPaths[name] = { name: name, scope: hookScope };
     };
-    /** * @hidden */
+    /** * @hidden */ // tslint:disable-next-line
     TransitionService.prototype._getPathTypes = function () {
         return this._criteriaPaths;
     };
@@ -6685,6 +6737,7 @@ var TransitionService = /** @class */ (function () {
         // Wire up Resolve hooks
         fns.eagerResolve = registerEagerResolvePath(this);
         fns.lazyResolve = registerLazyResolveState(this);
+        fns.resolveAll = registerResolveRemaining(this);
         // Wire up the View management hooks
         fns.loadViews = registerLoadEnteringViews(this);
         fns.activateViews = registerActivateViews(this);
@@ -6776,7 +6829,7 @@ var StateService = /** @class */ (function () {
     });
     /** @internalapi */
     StateService.prototype.dispose = function () {
-        this.defaultErrorHandler(noop$1);
+        this.defaultErrorHandler(noop);
         this.invalidCallbacks = [];
     };
     /**
@@ -7466,29 +7519,11 @@ function locationPluginFactory(name, isHtml5, serviceClass, configurationClass) 
         return { name: name, service: service, configuration: configuration, dispose: dispose };
     };
 }
-function getCustomEventCtor() {
-    // CustomEvent Polyfill
-    function _CustomEvent(event, params) {
-        params = params || { bubbles: false, cancelable: false, detail: undefined };
-        var evt = document.createEvent('CustomEvent');
-        evt.initCustomEvent(event, params.bubbles, params.cancelable, params.detail);
-        return evt;
-    }
-    _CustomEvent.prototype = Event.prototype;
-    try {
-        new CustomEvent('foo');
-        return CustomEvent;
-    }
-    catch (_err) {
-        return _CustomEvent;
-    }
-}
 
 /**
  * @internalapi
  * @module vanilla
  */ /** */
-var Evt = getCustomEventCtor();
 /** A base `LocationServices` */
 var BaseLocationServices = /** @class */ (function () {
     function BaseLocationServices(router, fireAfterUpdate) {
@@ -7499,16 +7534,15 @@ var BaseLocationServices = /** @class */ (function () {
         this.hash = function () { return parseUrl$1(_this._get()).hash; };
         this.path = function () { return parseUrl$1(_this._get()).path; };
         this.search = function () { return getParams(parseUrl$1(_this._get()).search); };
-        this._location = self && self.location;
-        this._history = self && self.history;
+        this._location = root.location;
+        this._history = root.history;
     }
     BaseLocationServices.prototype.url = function (url, replace) {
         if (replace === void 0) { replace = true; }
         if (isDefined(url) && url !== this._get()) {
             this._set(null, null, url, replace);
             if (this.fireAfterUpdate) {
-                var evt_1 = extend(new Evt("locationchange"), { url: url });
-                this._listeners.forEach(function (cb) { return cb(evt_1); });
+                this._listeners.forEach(function (cb) { return cb({ url: url }); });
             }
         }
         return buildUrl(this);
@@ -7544,7 +7578,7 @@ var HashLocationService = /** @class */ (function (_super) {
     __extends(HashLocationService, _super);
     function HashLocationService(router) {
         var _this = _super.call(this, router, false) || this;
-        self.addEventListener('hashchange', _this._listener, false);
+        root.addEventListener('hashchange', _this._listener, false);
         return _this;
     }
     HashLocationService.prototype._get = function () {
@@ -7555,7 +7589,7 @@ var HashLocationService = /** @class */ (function (_super) {
     };
     HashLocationService.prototype.dispose = function (router) {
         _super.prototype.dispose.call(this, router);
-        self.removeEventListener('hashchange', this._listener);
+        root.removeEventListener('hashchange', this._listener);
     };
     return HashLocationService;
 }(BaseLocationServices));
@@ -7610,7 +7644,7 @@ var PushStateLocationService = /** @class */ (function (_super) {
     function PushStateLocationService(router) {
         var _this = _super.call(this, router, true) || this;
         _this._config = router.urlService.config;
-        self.addEventListener("popstate", _this._listener, false);
+        root.addEventListener('popstate', _this._listener, false);
         return _this;
     }
     
@@ -7620,26 +7654,33 @@ var PushStateLocationService = /** @class */ (function (_super) {
      * - trailing filename
      * - protocol and hostname
      *
+     * If <base href='/base/'>, this returns '/base'.
+     * If <base href='/foo/base/'>, this returns '/foo/base'.
      * If <base href='/base/index.html'>, this returns '/base'.
      * If <base href='http://localhost:8080/base/index.html'>, this returns '/base'.
+     * If <base href='/base'>, this returns ''.
+     * If <base href='http://localhost:8080'>, this returns ''.
+     * If <base href='http://localhost:8080/'>, this returns ''.
      *
      * See: https://html.spec.whatwg.org/dev/semantics.html#the-base-element
      */
     PushStateLocationService.prototype._getBasePrefix = function () {
-        return stripFile(this._config.baseHref());
+        return stripLastPathElement(this._config.baseHref());
     };
     PushStateLocationService.prototype._get = function () {
         var _a = this._location, pathname = _a.pathname, hash = _a.hash, search = _a.search;
         search = splitQuery(search)[1]; // strip ? if found
         hash = splitHash(hash)[1]; // strip # if found
         var basePrefix = this._getBasePrefix();
-        var exactMatch = pathname === this._config.baseHref();
-        var startsWith = pathname.startsWith(basePrefix);
-        pathname = exactMatch ? '/' : startsWith ? pathname.substring(basePrefix.length) : pathname;
-        return pathname + (search ? "?" + search : "") + (hash ? "#" + hash : "");
+        var exactBaseHrefMatch = pathname === this._config.baseHref();
+        var startsWithBase = pathname.substr(0, basePrefix.length) === basePrefix;
+        pathname = exactBaseHrefMatch ? '/' : startsWithBase ? pathname.substring(basePrefix.length) : pathname;
+        return pathname + (search ? '?' + search : '') + (hash ? '#' + hash : '');
     };
     PushStateLocationService.prototype._set = function (state, title, url, replace) {
-        var fullUrl = this._getBasePrefix() + url;
+        var basePrefix = this._getBasePrefix();
+        var slash = url && url[0] !== '/' ? '/' : '';
+        var fullUrl = (url === '' || url === '/') ? this._config.baseHref() : basePrefix + slash + url;
         if (replace) {
             this._history.replaceState(state, title, fullUrl);
         }
@@ -7649,7 +7690,7 @@ var PushStateLocationService = /** @class */ (function (_super) {
     };
     PushStateLocationService.prototype.dispose = function (router) {
         _super.prototype.dispose.call(this, router);
-        self.removeEventListener("popstate", this._listener);
+        root.removeEventListener('popstate', this._listener);
     };
     return PushStateLocationService;
 }(BaseLocationServices));
@@ -7669,7 +7710,7 @@ var MemoryLocationConfig = /** @class */ (function () {
         this.baseHref = function () { return _this._baseHref; };
         this.html5Mode = function () { return false; };
         this.hashPrefix = function (newval) { return isDefined(newval) ? _this._hashPrefix = newval : _this._hashPrefix; };
-        this.dispose = noop$1;
+        this.dispose = noop;
     }
     return MemoryLocationConfig;
 }());
@@ -7712,7 +7753,7 @@ var BrowserLocationConfig = /** @class */ (function () {
     };
     BrowserLocationConfig.prototype.applyDocumentBaseHref = function () {
         var baseTag = document.getElementsByTagName("base")[0];
-        return this._baseHref = baseTag ? baseTag.href.substr(location.origin.length) : "";
+        return this._baseHref = baseTag ? baseTag.href.substr(location.origin.length) : location.pathname || '/';
     };
     BrowserLocationConfig.prototype.dispose = function () { };
     return BrowserLocationConfig;
@@ -7767,13 +7808,14 @@ var UIRouterPluginBase = /** @class */ (function () {
 
 
 var index$1 = Object.freeze({
+	root: root,
 	fromJson: fromJson,
 	toJson: toJson,
 	forEach: forEach,
 	extend: extend,
 	equals: equals,
 	identity: identity,
-	noop: noop$1,
+	noop: noop,
 	createProxyFunctions: createProxyFunctions,
 	inherit: inherit,
 	inArray: inArray,
@@ -7855,7 +7897,7 @@ var index$1 = Object.freeze({
 	stringify: stringify,
 	beforeAfterSubstr: beforeAfterSubstr,
 	hostRegex: hostRegex,
-	stripFile: stripFile,
+	stripLastPathElement: stripLastPathElement,
 	splitHash: splitHash,
 	splitQuery: splitQuery,
 	splitEqual: splitEqual,
@@ -7920,7 +7962,6 @@ var index$1 = Object.freeze({
 	parseUrl: parseUrl$1,
 	buildUrl: buildUrl,
 	locationPluginFactory: locationPluginFactory,
-	getCustomEventCtor: getCustomEventCtor,
 	servicesPlugin: servicesPlugin,
 	hashLocationPlugin: hashLocationPlugin,
 	pushStateLocationPlugin: pushStateLocationPlugin,
@@ -8370,13 +8411,18 @@ var getStateHookBuilder = function (hookName) {
         var pathname = hookName === 'onExit' ? 'from' : 'to';
         function decoratedNg1Hook(trans, state) {
             var resolveContext = new ResolveContext(trans.treeChanges(pathname));
-            var locals = extend(getLocals(resolveContext), { $state$: state, $transition$: trans });
+            var subContext = resolveContext.subContext(state.$$state());
+            var locals = extend(getLocals(subContext), { $state$: state, $transition$: trans });
             return services.$injector.invoke(hook, this, locals);
         }
         return hook ? decoratedNg1Hook : undefined;
     };
 };
 
+/**
+ * @internalapi
+ * @module ng1
+ */ /** */
 /**
  * Implements UI-Router LocationServices and LocationConfig using Angular 1's $location service
  */
@@ -8401,7 +8447,7 @@ var Ng1LocationServices = /** @class */ (function () {
     };
     Ng1LocationServices.prototype.url = function (newUrl, replace, state) {
         if (replace === void 0) { replace = false; }
-        if (newUrl)
+        if (isDefined(newUrl))
             this.$location.url(newUrl);
         if (replace)
             this.$location.replace();
@@ -9236,7 +9282,7 @@ uiState = ['$uiRouter', '$timeout',
                 var rawDef = {};
                 var getDef = function () { return processedDef($state, element, rawDef); };
                 var inputAttrs = ['uiState', 'uiStateParams', 'uiStateOpts'];
-                var watchDeregFns = inputAttrs.reduce(function (acc, attr) { return (acc[attr] = noop$1, acc); }, {});
+                var watchDeregFns = inputAttrs.reduce(function (acc, attr) { return (acc[attr] = noop, acc); }, {});
                 function update() {
                     var def = getDef();
                     if (unlinkInfoFn)
@@ -9337,7 +9383,7 @@ uiState = ['$uiRouter', '$timeout',
  * </div>
  * ```
  *
- * When the current state is "admin.roles" the "active" class will be applied to both the <div> and <a> elements.
+ * When the current state is "admin.roles" the "active" class will be applied to both the `<div>` and `<a>` elements.
  * It is important to note that the state names/globs passed to `ui-sref-active` override any state provided by a linked `ui-sref`.
  *
  * ### Notes:
@@ -9387,7 +9433,7 @@ uiSrefActive = ['$state', '$stateParams', '$interpolate', '$uiRouter',
                         return deregister;
                     };
                     function updateAfterTransition(trans) {
-                        trans.promise.then(update, noop$1);
+                        trans.promise.then(update, noop);
                     }
                     $scope.$on('$stateChangeSuccess', update);
                     $scope.$on('$destroy', $uiRouter.transitionService.onStart({}, updateAfterTransition));
@@ -9774,7 +9820,7 @@ function $ViewDirectiveFill($compile, $controller, $transitions, $view, $q, $tim
                     $compile($element.contents())(scope);
                     return;
                 }
-                var cfg = data.$cfg || { viewDecl: {}, getTemplate: ng_from_import.noop };
+                var cfg = data.$cfg || { viewDecl: {}, getTemplate: noop };
                 var resolveCtx = cfg.path && new ResolveContext(cfg.path);
                 $element.html(cfg.getTemplate($element, resolveCtx) || initial);
                 trace.traceUIViewFill(data.$uiView, $element.html());
@@ -9920,13 +9966,14 @@ exports.ng1ViewsBuilder = ng1ViewsBuilder;
 exports.Ng1ViewConfig = Ng1ViewConfig;
 exports.StateProvider = StateProvider;
 exports.UrlRouterProvider = UrlRouterProvider;
+exports.root = root;
 exports.fromJson = fromJson;
 exports.toJson = toJson;
 exports.forEach = forEach;
 exports.extend = extend;
 exports.equals = equals;
 exports.identity = identity;
-exports.noop = noop$1;
+exports.noop = noop;
 exports.createProxyFunctions = createProxyFunctions;
 exports.inherit = inherit;
 exports.inArray = inArray;
@@ -10008,7 +10055,7 @@ exports.fnToString = fnToString;
 exports.stringify = stringify;
 exports.beforeAfterSubstr = beforeAfterSubstr;
 exports.hostRegex = hostRegex;
-exports.stripFile = stripFile;
+exports.stripLastPathElement = stripLastPathElement;
 exports.splitHash = splitHash;
 exports.splitQuery = splitQuery;
 exports.splitEqual = splitEqual;
@@ -10068,7 +10115,6 @@ exports.getParams = getParams;
 exports.parseUrl = parseUrl$1;
 exports.buildUrl = buildUrl;
 exports.locationPluginFactory = locationPluginFactory;
-exports.getCustomEventCtor = getCustomEventCtor;
 exports.servicesPlugin = servicesPlugin;
 exports.hashLocationPlugin = hashLocationPlugin;
 exports.pushStateLocationPlugin = pushStateLocationPlugin;